+++ .envrc
@@ -1,2 +1,2 @@
 export PROJ_NAME="VideoFilter"
-export VIDEO_NAME="skb"
+export VIDEO_NAME="soccer"
