+++ CMakeLists.txt
@@ -13,6 +13,7 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 set(SOURCES
     src/platform/linux/display_sdl.cc
+    src/platform/linux/pl_mpeg.cc
     src/common/filter.cc
 )
